+++ curr_scraper/apps/api_caller/api/v1/views.py
@@ -1,15 +1,12 @@
 import django_filters
 from api_caller.models import Currency, CurrencyRateRecord
-from django.db import connection, OperationalError
 from django.http import HttpResponse
 from django_filters import filters
 from django_filters.rest_framework import DjangoFilterBackend
-from rest_framework import status, generics
+from rest_framework import viewsets
 from rest_framework.pagination import PageNumberPagination
-from rest_framework.response import Response
-from rest_framework.views import APIView
 
-from .serializers import CurrencySerializer, CurrencyRateRecordSerializer
+from curr_scraper.apps.api_caller.api.v1.serializers import CurrencySerializer, CurrencyRateRecordSerializer
 
 
 def index(request):
@@ -17,20 +14,7 @@ def index(request):
     return HttpResponse("Hello!")
 
 
-class CheckHealth(APIView):
-    """Return "OK" if server is up."""
-    def get(self, request):
-        try:
-            connection.ensure_connection()
-        except OperationalError:
-            content = {'Server status': 'Failed connection to database'}
-            return Response(content, status=status.HTTP_503_SERVICE_UNAVAILABLE)
-        else:
-            content = {'Server status': 'OK'}
-            return Response(content, status=status.HTTP_200_OK)
-
-
-class CurrencyView(generics.ListAPIView):
+class CurrencyView(viewsets.ModelViewSet):
     """Get a list of all currencies."""
     queryset = Currency.objects.all()
     serializer_class = CurrencySerializer
@@ -56,7 +40,7 @@ class RatesFilter(django_filters.FilterSet):
         fields = ('curr',)
 
 
-class CurrencyRateView(generics.ListAPIView):
+class CurrencyRateView(viewsets.ModelViewSet):
     """Get a list of all currency rate records with optional filters."""
     queryset = CurrencyRateRecord.objects.all()
     filter_backends = [DjangoFilterBackend]
